Work around Gnuplot bug
In newer versions of Gnuplot, the ylabel can go missing in the dumb
terminal.  This appears to be because the centring of the label text
isn't correctly applied and the label protrudes outside the graph area.
Gnuplot then refuses to draw it.   The work-around is to apply a
horizontal offset equal to half the length of the label string (rounded
down).

Fixes #18.