Make this command-line tool not entirely broken
What does this tool even do?  I suspect it probably relates to past
investigations of instability in the monochromator on beamline I19 of
Diamond Light Source.  Does it still work?  It does something.  Does it
do that thing well?  It's not clear.  Is it tested?  No.  Does it have
command line arguments that don't do anything?  Yes.

Looks like firewood to me, but for now, let's just make it sort-of work.