Remove stability_fft.py and related test and utils (#27)
What does this tool even do?
I suspect it probably relates to past investigations of instability in the monochromator on beamline I19 of Diamond Light Source. It certainly has nothing to do with screen19 and no-one seems to use it.

Does it do that thing well?
It's not clear. It has at least one command line parameter that does nothing.

Does it still work?
It does now (I just fixed it in ee5c4f3). For a long time, it was simply broken and it seems no-one noticed.

Is it tested?
Sort of — there are unit tests of two utilities (one of which is not used anywhere).

Docs?
Code is documentation, right?

So?
Remove it.