Remove fast_mode, do French Wilson scaling (#29)
Previously, if screen19 was called on a sweep of fewer than ten images, 
it ran in fast_mode, which bypassed profile modelling (and also
dials.report).  However, this is a poor heuristic for determining that 
profile modelling will fail.  In fact, it is usually possible to create
an adequate profile model from a single image.  In any case, we can 
default to using summation intensities for the Wilson plot analysis if 
profile modelling fails.

Also introduce French-Wilson scaling of the integrated intensities, to 
boost the number of valid reflections.  Previously, reflections with 
negative intensity were simply discarded.  This is probably because I 
wrote this code before I knew anything much about crystallography and 
had never heard of French-Wilson scaling.  It seems much more sensible 
to use it.